# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BarberMetrics API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Appointment {
  id: number;
  date: string;
  startTime: string;
  endTime: string;
  durationMinutes: number;
  service: string;
  value: number;
  barberEarnings: number;
  createdAt: string;
}

export interface CreateAppointmentBody {
  date?: string;
  startTime?: string;
  endTime?: string;
  durationMinutes: number;
  service: string;
  value: number;
}

export interface UpdateAppointmentBody {
  date?: string;
  startTime?: string;
  endTime?: string;
  service?: string;
  value?: number;
}

export interface CommissionSetting {
  /**
   * @minimum 0
   * @maximum 100
   */
  commissionPercent: number;
}

export interface TimerSession {
  id: string;
  startedAt: string;
  elapsedSeconds: number;
  isActive: boolean;
}

export interface FinishTimerBody {
  service: string;
  value: number;
  customService?: string;
}

export interface Bill {
  id: number;
  name: string;
  value: number;
  dueDay: number;
  category?: string;
  createdAt: string;
}

export interface CreateBillBody {
  name: string;
  value: number;
  dueDay: number;
  category?: string;
}

export interface DashboardSummary {
  clientsToday: number;
  grossRevenue: number;
  barberEarnings: number;
  avgDurationMinutes: number;
  earningsPerHour: number;
  dailyGoal: number;
  goalProgress: number;
  isTimerActive: boolean;
  timerStartedAt?: string | null;
  minimumDailyGoal: number;
  chairGoal: number;
  productivityPercent: number;
  chairValuePerHour: number;
  barberEarningsPerHour: number;
  totalWorkingMinutes: number;
  totalServiceMinutes: number;
}

export interface ServiceBreakdown {
  service: string;
  count: number;
  percentage: number;
  totalRevenue: number;
  avgValue: number;
}

export interface ProductivityStats {
  period: string;
  totalClients: number;
  avgDurationMinutes: number;
  earningsPerHour: number;
  totalWorkingMinutes: number;
  totalServiceMinutes: number;
  idleMinutes: number;
  potentialExtraEarnings: number;
  serviceBreakdown: ServiceBreakdown[];
  dailyAvgClients: number;
  totalEarnings: number;
  grossRevenue: number;
  productivityPercent: number;
  chairValuePerHour: number;
  barberEarningsPerHour: number;
  chairGoal: number;
  minimumDailyGoal: number;
}

export type ProductivityTipType =
  (typeof ProductivityTipType)[keyof typeof ProductivityTipType];

export const ProductivityTipType = {
  time: "time",
  revenue: "revenue",
  service: "service",
  idle: "idle",
} as const;

export interface ProductivityTip {
  id: string;
  type: ProductivityTipType;
  message: string;
  impact?: string | null;
}

export interface FinancialSummary {
  currentMonthEarnings: number;
  totalBills: number;
  remainingAfterBills: number;
  billsShortfall: number;
  weeklyGoal: number;
  dailyGoalRequired: number;
  projectedMonthlyEarnings: number;
  projectedAfterBills: number;
  daysRemainingInMonth: number;
  billsList: Bill[];
  onTrack: boolean;
  dailyDeficit?: number | null;
}

export interface DailyGoal {
  goal: number;
}

export interface UpdateDailyGoalBody {
  goal: number;
}

export interface WorkHours {
  hoursPerDay: number;
  daysPerWeek: number;
}

export interface DayEarnings {
  date: string;
  earnings: number;
  appointmentCount: number;
}

export interface HourStat {
  hour: number;
  count: number;
}

export interface WeekdayStat {
  weekday: number;
  dayName: string;
  avgEarnings: number;
  count: number;
}

export interface ServiceRankItem {
  service: string;
  count: number;
  totalEarnings: number;
}

export interface MonthlyAnalysis {
  month: string;
  workedDays: number;
  totalEarnings: number;
  dailyAverage: number;
  monthlyForecast: number;
  topDays: DayEarnings[];
  busiestHours: HourStat[];
  weekdayStats: WeekdayStat[];
  serviceRanking: ServiceRankItem[];
}

export type ListAppointmentsParams = {
  period?: ListAppointmentsPeriod;
};

export type ListAppointmentsPeriod =
  (typeof ListAppointmentsPeriod)[keyof typeof ListAppointmentsPeriod];

export const ListAppointmentsPeriod = {
  today: "today",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetMonthlyAnalysisParams = {
  /**
   * Month in YYYY-MM format, defaults to current month
   */
  month?: string;
};

export type GetProductivityStatsParams = {
  period?: GetProductivityStatsPeriod;
};

export type GetProductivityStatsPeriod =
  (typeof GetProductivityStatsPeriod)[keyof typeof GetProductivityStatsPeriod];

export const GetProductivityStatsPeriod = {
  today: "today",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetProductivityTipsParams = {
  period?: GetProductivityTipsPeriod;
};

export type GetProductivityTipsPeriod =
  (typeof GetProductivityTipsPeriod)[keyof typeof GetProductivityTipsPeriod];

export const GetProductivityTipsPeriod = {
  today: "today",
  week: "week",
  month: "month",
  year: "year",
} as const;

export type GetStatementParams = {
  start: string;
  end: string;
};