# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BarberMetrics API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
            default: today
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List monthly bills
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillBody"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    delete:
      operationId: deleteBill
      tags: [bills]
      summary: Delete a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillBody"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        durationMinutes:
          type: integer
        service:
          type: string
        value:
          type: number
        barberEarnings:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - startTime
        - endTime
        - durationMinutes
        - service
        - value
        - barberEarnings
        - createdAt
    CreateAppointmentBody:
      type: object
      properties:
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        durationMinutes:
          type: integer
        service:
          type: string
        value:
          type: number
      required:
        - service
        - value
        - durationMinutes
    UpdateAppointmentBody:
      type: object
      properties:
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        service:
          type: string
        value:
          type: number
...
    Bill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        value:
          type: number
        dueDay:
          type: integer
        category:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - value
        - dueDay
        - createdAt
    CreateBillBody:
      type: object
      properties:
        name:
          type: string
        value:
          type: number
        dueDay:
          type: integer
        category:
          type: string
      required:
        - name
        - value
        - dueDay
...